Controls/requirements.txt
Click==7.0
pip==20.0.2
pyserial==3.4
PySide2==5.14.1
python-dotenv==0.11.0
setuptools==45.2.0
shiboken2==5.14.1
